Separate condition handlers from api file
